更新博客前后端链条相关流程

+++ src/main/java/com/qxc/blog/service/impl/EditBlogServiceImpl.java
@@ -1,12 +1,9 @@
 package com.qxc.blog.service.impl;
 
-import com.qxc.blog.AOPInterceptor.BlogEvent.BlogEvent;
-import com.qxc.blog.AOPInterceptor.BlogEvent.BlogEventPublisher;
-import com.qxc.blog.AOPInterceptor.LogRecod.LogRecord;
-import com.qxc.blog.pojo.Blog;
-import com.qxc.blog.pojo.BlogComments;
-import com.qxc.blog.pojo.BlogContent;
-import com.qxc.blog.pojo.BlogRelationship;
+import com.qxc.blog.aopInterceptor.aop.LogRecod.LogRecord;
+import com.qxc.blog.aopInterceptor.event.blogEvent.BlogEvent;
+import com.qxc.blog.aopInterceptor.event.blogEvent.BlogEventPublisher;
+import com.qxc.blog.pojo.*;
 import com.qxc.blog.self.BlogAndContent;
 import com.qxc.blog.self.BlogEventEnum;
 import com.qxc.blog.service.*;
@@ -42,6 +39,9 @@ public class EditBlogServiceImpl implements EditBlogService {
     @Resource
     private BlogRelationshipService blogRelationshipService;
 
+    @Resource
+    private BlogAccessService blogAccessService;
+
     /**
      * 获取整个博客全部信息
      *
@@ -49,7 +49,10 @@ public class EditBlogServiceImpl implements EditBlogService {
      * @return
      */
     @Override
-    public BlogAndContent getBlogById(String articalId) {
+    public BlogAndContent getBlogById(BlogUser blogUser, String articalId) {
+        if (!blogAccessService.checkSee(blogUser, articalId)) {
+            return null;
+        }
         final Blog blog = blogService.getBlogById(articalId);
         if (Objects.isNull(blog)) {
             return null;
@@ -78,7 +81,10 @@ public class EditBlogServiceImpl implements EditBlogService {
     @Override
     @Transactional
     @LogRecord(param = true, time = true)
-    public boolean addBlog(@NotNull BlogAndContent blogAndContent) throws Exception {
+    public boolean addBlog(BlogUser blogUser, @NotNull BlogAndContent blogAndContent) throws Exception {
+        if (!blogAccessService.checkAdd(blogUser, blogAndContent.getBlogRelationship().getFirst().getFatheridx())) {
+            throw new Exception("add blog failed");
+        }
         if (!blogService.addBlog(blogAndContent.getBlog())) {
             throw new Exception("add blog failed");
         }
@@ -94,6 +100,31 @@ public class EditBlogServiceImpl implements EditBlogService {
     }
 
     /**
+     * 创建新博客
+     *
+     * @param blogAndContent
+     * @return
+     */
+    @Override
+    @Transactional
+    @LogRecord(param = true, time = true)
+    public boolean addBlogWithOutCheck(BlogUser blogUser, @NotNull BlogAndContent blogAndContent) throws Exception {
+        if (!blogService.addBlog(blogAndContent.getBlog())) {
+            throw new Exception("add blog failed");
+        }
+        if (!blogContentService.addBlogContent(blogAndContent.getBlogContent())) {
+            throw new Exception("add blog content failed");
+        }
+        if (!blogRelationshipService.addBlogRelationship(blogAndContent.getBlogRelationship().getFirst())) {
+            throw new Exception("add blog relationship failed");
+        }
+        // 发布事件els可能订阅
+        blogAddEventPublisher.publish(new BlogEvent(blogAndContent.getBlog().getArticleid(), BlogEventEnum.CREATE));
+        return true;
+    }
+
+
+    /**
      * 更新博客
      *
      * @param blogAndContent
@@ -102,7 +133,10 @@ public class EditBlogServiceImpl implements EditBlogService {
     @Override
     @Transactional
     @LogRecord(param = true, time = true)
-    public boolean updateBlog(@NotNull BlogAndContent blogAndContent) throws Exception {
+    public boolean updateBlog(BlogUser blogUser, @NotNull BlogAndContent blogAndContent) throws Exception {
+        if (!blogAccessService.checkUpdate(blogUser, blogAndContent.getBlog().getArticleid())) {
+            return false;
+        }
         if (!blogService.updateBlog(blogAndContent.getBlog())) {
             throw new Exception("add blog failed");
         }
@@ -123,7 +157,10 @@ public class EditBlogServiceImpl implements EditBlogService {
     @Override
     @Transactional
     @LogRecord(param = true, time = true)
-    public boolean deleteBlog(@NotNull BlogAndContent blogAndContent) throws Exception {
+    public boolean deleteBlog(BlogUser blogUser, @NotNull BlogAndContent blogAndContent) throws Exception {
+        if (!blogAccessService.checkDelete(blogUser, blogAndContent.getBlog().getArticleid())) {
+            return false;
+        }
         final List<String> relationship = blogRelationshipService.deleteBlogRelationship(blogAndContent.getBlog().getArticleid());
         if (!blogService.deleteBlog(blogAndContent.getBlog()) || !blogService.deleteBlog(relationship)) {
             throw new Exception("delete blog failed");
